trying to maker SpatialDataContainer.java cloud-compatible; many things still missing (added TODO's)

+++ src/main/java/io/SpatialDataContainer.java
@@ -5,8 +5,6 @@ import util.Cloud;
 
 import org.janelia.saalfeldlab.n5.DataType;
 import org.janelia.saalfeldlab.n5.GzipCompression;
-import org.janelia.saalfeldlab.n5.N5FSReader;
-import org.janelia.saalfeldlab.n5.N5FSWriter;
 
 import java.io.File;
 import java.io.IOException;
@@ -32,6 +30,7 @@ import org.janelia.saalfeldlab.n5.universe.N5Factory.StorageFormat;
 public class SpatialDataContainer {
 
 	final private String rootPath;
+	final private URI rootPathURI;
 	final private boolean readOnly;
 	final private ExecutorService service;
 	final private N5Reader n5;
@@ -49,6 +48,7 @@ public class SpatialDataContainer {
 
 	protected SpatialDataContainer(final String path, final ExecutorService service, final boolean readOnly) throws IOException {
 		this.rootPath = path;
+		this.rootPathURI = URI.create( this.rootPath );
 		this.readOnly = readOnly;
 		this.service = service;
 
@@ -95,7 +95,7 @@ public class SpatialDataContainer {
 	}
 
 	protected void initializeContainer() {
-		N5FSWriter writer = (N5FSWriter) n5;
+		N5Writer writer = (N5Writer) n5;
 		writer.setAttribute("/", versionKey, version);
 		writer.createGroup("/matches");
 		updateDatasetMetadata();
@@ -123,7 +123,7 @@ public class SpatialDataContainer {
 	}
 
 	protected void updateDatasetMetadata() {
-		N5FSWriter writer = (N5FSWriter) n5;
+		N5Writer writer = (N5Writer) n5;
 		writer.setAttribute("/", numDatasetsKey, datasets.size());
 		writer.setAttribute("/", datasetsKey, datasets.toArray());
 	}
@@ -133,8 +133,13 @@ public class SpatialDataContainer {
 	}
 
 	public void addExistingDataset(String path, String locationPath, String exprValuePath, String annotationPath, String geneAnnotationPath) {
-		associateDataset(path, (src, dest) -> {
-			try {return Files.move(src, dest);}
+		associateDataset(path, (src, dest) ->
+		{
+			try
+			{
+				// TODO: not cloud compatible yet
+				return Files.move(src, dest);
+			}
 			catch (IOException e) {throw new SpatialDataException("Could not move dataset to container.", e);}
 		}, locationPath, exprValuePath, annotationPath, geneAnnotationPath);
 	}
@@ -144,8 +149,13 @@ public class SpatialDataContainer {
 	}
 
 	public void linkExistingDataset(String path, String locationPath, String exprValuePath, String annotationPath, String geneAnnotationPath) {
-		associateDataset(path, (target, link) -> {
-			try {return Files.createSymbolicLink(link, target);}
+		associateDataset(path, (target, link) ->
+		{
+			try
+			{
+				// TODO: not cloud compatible yet
+				return Files.createSymbolicLink(link, target);
+			}
 			catch (IOException e) {throw new SpatialDataException("Could not link dataset to container.", e);}
 		}, locationPath, exprValuePath, annotationPath, geneAnnotationPath);
 	}
@@ -158,6 +168,7 @@ public class SpatialDataContainer {
 			String annotationPath,
 			String geneAnnotationPath) {
 
+		// TODO: not cloud compatible yet
 		Path oldPath = Paths.get(path);
 		if (!oldPath.toFile().exists()) {
 			throw new IllegalArgumentException("Dataset '" + oldPath + "' does not exist.");
@@ -184,11 +195,19 @@ public class SpatialDataContainer {
 			writer.setAttribute("/", datasetName + geneAnnotationPathKey, geneAnnotationPath);
 	}
 
-	public void deleteDataset(String datasetName) throws IOException {
+	public void deleteDataset(String datasetName) throws IOException
+	{
 		if (readOnly)
+		{
 			throw new IllegalStateException("Trying to modify a read-only spatial data container.");
-		if (datasets.remove(datasetName)) {
-			deleteFileOrDirectory(Paths.get(rootPath, datasetName));
+		}
+		if (datasets.remove(datasetName))
+		{
+			// TODO: no cloud support yet
+			if ( Cloud.isFile( rootPathURI ))
+				deleteFileOrDirectory(Paths.get(rootPath, datasetName));
+			else
+				throw new RuntimeException( "not supported for cloud yet." );
 			updateDatasetMetadata();
 		}
 	}
@@ -202,7 +221,14 @@ public class SpatialDataContainer {
 		String path2 = n5.getAttribute("/", datasetName + exprValuePathKey, String.class);
 		String path3 = n5.getAttribute("/", datasetName + annotationPathKey, String.class);
 		String path4 = n5.getAttribute("/", datasetName + geneAnnotationPathKey, String.class);
-		SpatialDataIO sdio = SpatialDataIO.open(Paths.get(rootPath, datasetName).toRealPath().toString(), service);
+
+		final String path;
+		if ( Cloud.isFile( rootPathURI ))
+			path = Paths.get(rootPath, datasetName).toRealPath().toString();
+		else
+			path = Cloud.appendName(rootPathURI, datasetName);
+
+		SpatialDataIO sdio = SpatialDataIO.open( path, service);
 		sdio.setDataPaths(path1, path2, path3, path4);
 		return sdio;
 	}
@@ -214,7 +240,14 @@ public class SpatialDataContainer {
 		String path2 = n5.getAttribute("/", datasetName + exprValuePathKey, String.class);
 		String path3 = n5.getAttribute("/", datasetName + annotationPathKey, String.class);
 		String path4 = n5.getAttribute("/", datasetName + geneAnnotationPathKey, String.class);
-		SpatialDataIO sdio = SpatialDataIO.openReadOnly(Paths.get(rootPath, datasetName).toRealPath().toString(), service);
+
+		final String path;
+		if ( Cloud.isFile( rootPathURI ))
+			path = Paths.get(rootPath, datasetName).toRealPath().toString();
+		else
+			path = Cloud.appendName(rootPathURI, datasetName);
+
+		SpatialDataIO sdio = SpatialDataIO.openReadOnly(path, service);
 		sdio.setDataPaths(path1, path2, path3, path4);
 		return sdio;
 	}
@@ -259,7 +292,12 @@ public class SpatialDataContainer {
 			deleteFileOrDirectory(Paths.get(rootPath, "matches", matchName));
 	}
 
-	public void deleteFileOrDirectory(Path path) throws IOException {
+	public void deleteFileOrDirectory(Path path) throws IOException
+	{
+		// TODO: no cloud support yet
+		if ( Cloud.isFile( rootPathURI ))
+			throw new RuntimeException( "not supported for cloud yet." );
+
 		File file = new File(path.toString());
 		if (file.exists()) {
 			if (file.isFile())
@@ -270,7 +308,7 @@ public class SpatialDataContainer {
 	}
 
 	public void savePairwiseMatch(final SiftMatch results) {
-		N5FSWriter writer = (N5FSWriter) n5;
+		N5Writer writer = (N5Writer) n5;
 		final String matchName = constructMatchName(results.getStDataAName(), results.getStDataBName());
 		final String pairwiseGroupName = writer.groupPath("/", "matches", matchName);
 
@@ -332,6 +370,7 @@ public class SpatialDataContainer {
 			return stDataBName + "-" + stDataAName;
 	}
 
+	// TODO: no cloud support yet
 	private static class TreeDeleter extends SimpleFileVisitor<Path> {
 		@Override
 		public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
